Refs #3137: Used coord_t and signal_t in appropriate spots in MD* classes. Right now both are typedef'd to double but it builds if typedef'd so something else e.g float.

+++ Code/Mantid/Framework/Nexus/src/LoadNexusLogs.cpp
@@ -343,7 +343,7 @@ namespace Mantid
       // Now the actual data
       ::NeXus::Info info = file.getInfo();
       // Check the size
-      if( info.dims[0] != time_double.size() )
+      if( size_t(info.dims[0]) != time_double.size() )
       {
         file.closeData();
         throw ::NeXus::Exception("Invalid value entry for time series");
